Added `ser/sets` module
